rename : package name (pop-up-modal -> pop-up)
