+++ app/running/finish.tsx
@@ -13,7 +13,7 @@ import { useNavigation, useRoute } from '@react-navigation/native';
 import EmptyHeartIcon from '@/assets/icons/emptyHeart.svg';
 import { CourseButton } from '@/components/button/RunningButton';
 import { BackHeader } from '@/components/header/IconHeader';
-import { MyCourseSaveModal } from '@/components/modal/pop-up-modal/CourseModal';
+import { MyCourseSaveModal } from '@/components/modal/pop-up/CourseModal';
 import Colors from '@/constants/Colors';
 import Sizes from '@/constants/Sizes';
 import {
